git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@10972 01821946-851b-0410-b7f4-f1477c70e57a

+++ 2d/palette.cpp
@@ -463,7 +463,7 @@ if (!i) {
 	Error(TXT_PAL_FILES, pszFile, DEFAULT_LEVEL_PALETTE);
 	return NULL;
 	}
-fsize	= cf.Length ();
+fsize	= (int) cf.Length ();
 Assert (fsize == 9472);
 palette.Read (cf);
 cf.Read (m_data.fadeTable, sizeof (m_data.fadeTable), 1);
